+++ src/components/budget/Budget.tsx
@@ -10,7 +10,6 @@ import {ProjectDataBudget} from '../../models/project';
 import {trendingUpOutline, pricetagOutline} from 'ionicons/icons';
 
 import {formatCurrency} from '../../utils/utils.currency';
-import {budgetRatio} from '../../utils/utils.budget';
 
 interface BudgetProps extends RootProps {
   budget: ProjectDataBudget | undefined;
@@ -48,24 +47,8 @@ const Budget: React.FC<BudgetProps> = (props) => {
     return (
       <IonLabel>
         <IonIcon icon={trendingUpOutline} aria-label={t('details.billed')} /> {formatCurrency(billed, props.settings.currency.currency)}
-        &nbsp;
-        {renderBudgetUsed(billed)}
       </IonLabel>
     );
-  }
-
-  function renderBudgetUsed(billed: number | undefined) {
-    if (!props.budget) {
-      return undefined;
-    }
-
-    const used: string | undefined = budgetRatio(props.budget.budget, billed);
-
-    if (!used) {
-      return undefined;
-    }
-
-    return <small>({used})</small>;
   }
 };